+++ src/controllers/sample/sample.controller.js
@@ -3,7 +3,7 @@ const {
   errorResponse,
 } = require('../../utils/responseHandler');
 
-const { statusCodes } = require('../../utils/statusCodes');
+const { statusCodes } = require('../../utils/statusCode');
 const { logger } = require('../../utils/logger');
 
 const sampleTest = async (req, res, next) => {
